move and expand validate all ingest templates test
